+ add handling player list with info tile on rack and points + fixed PlayerPoints class + fixed ClientConnection function + add static variable SkycrabGame

+++ SkyCrab/Common classes/Games/Players/PlayerPoints.cs
@@ -5,6 +5,6 @@ namespace SkyCrab.Common_classes.Games.Players
     public struct PlayerPoints
     {
         public UInt32 playerId;
-        public Int32 points;
+        public Int16 points;
     }
 }
